classification/requirements.txt
numpy
opencv_contrib_python
Pillow
tensorboardX
torch
tqdm
einops